Fix warning with endian integers

+++ include/bitstream/utility/endian.h
@@ -91,9 +91,9 @@ namespace bitstream::utility
     inline uint32_t endian_swap16(uint32_t value)
     {
 #if defined(_WIN32)
-        return _byteswap_ushort(value);
+        return _byteswap_ushort(static_cast<uint16_t>(value));
 #elif defined(__linux__)
-        return __builtin_bswap16(value);
+        return __builtin_bswap16(static_cast<uint16_t>(value));
 #else
         const uint32_t first = (value << 8) & 0x0000FF00;
         const uint32_t second = (value >> 8) & 0x000000FF;
